[fix] products manual

+++ database/migrations/2025_06_09_165703_add_device_id_to_orders_table.php
@@ -9,18 +9,24 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-public function up(): void
-{
-    Schema::table('orders', function (Blueprint $table) {
-        $table->string('device_id')->nullable()->after('due');
-    });
-}
+    public function up(): void
+    {
+        if (!Schema::hasColumn('orders', 'device_id')) {
+            Schema::table('orders', function (Blueprint $table) {
+                $table->string('device_id')->nullable()->after('due');
+            });
+        }
+    }
 
-public function down(): void
-{
-    Schema::table('orders', function (Blueprint $table) {
-        $table->dropColumn('device_id');
-    });
-}
-
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        if (Schema::hasColumn('orders', 'device_id')) {
+            Schema::table('orders', function (Blueprint $table) {
+                $table->dropColumn('device_id');
+            });
+        }
+    }
 };
